Fixes for quickstart.sh from Michael Add .hgignore

+++ docs/samples/quickstart.sh
@@ -1,5 +1,5 @@
-#!/bin/bash
-
+#!/bin/sh
+set -e
 # This script takes as arguments
 # 1) A tomcat distribution.
 # 
@@ -15,7 +15,22 @@
 DEFAULT_LOG_LEVEL=ERROR
 START_IN_BACKGROUND=FALSE
 
-while getopts ":vs" opt; do
+die() {
+  echo "$1" 2>&1
+  exit 1
+}
+
+ensuredir() {
+  [ "$1" ] || die "Empty directory name"
+  [ -d "$1" ] || install -d "$1"
+}
+
+# find the directory containing this script
+SRCDIR="$(dirname "$(readlink -f "$0")")"
+[ -r "$SRCDIR/mgmt.war" ] || die "Incorrect SRCDIR=$SRCDIR"
+DATADIR="$PWD"
+
+while getopts ":vsd:" opt; do
   case $opt in
     v)
       echo "Setting log levels to DEBUG." >&2
@@ -24,6 +39,9 @@ while getopts ":vs" opt; do
     s)
       echo "Starting in the background" >&2
       START_IN_BACKGROUND=TRUE
+      ;;
+    d)
+      DATADIR="$OPTARG"
       ;;
     \?)
       echo "Invalid option: -$OPTARG" >&2
@@ -34,31 +52,33 @@ done
 
 shift $(($OPTIND - 1))
 
-if [[ $# < 1 ]] 
-then
-  echo "Usage: $0 <Tomcat Distribution>"
-  exit 1
-fi
+[ $# -lt 1 ] && die "Usage: $0 <Tomcat Distribution>"
 
 TOMCAT_DISTRIBUTION=$1
 
-if [[ ! -f ${TOMCAT_DISTRIBUTION} ]]
-then
-  echo "${TOMCAT_DISTRIBUTION} is not a valid file?"
-  exit 1
-fi
+[ -f ${TOMCAT_DISTRIBUTION} ] || die "${TOMCAT_DISTRIBUTION} is not a valid file?"
 
 FQ_HOSTNAME=`hostname -f`
 
-if [[ -z ${FQ_HOSTNAME} ]]
+if [ -z "${FQ_HOSTNAME}" ]
 then
   echo "Unknown/empty hostname. Setting hostname to localhost"
   FQ_HOSTNAME="localhost"
 fi
 
+ensuredir "$DATADIR"
+
+cd "$DATADIR"
+DATADIR="$PWD"
+
+echo "SRCDIR=$SRCDIR"
+echo "DATADIR=$DATADIR"
+ 
+export ARCHAPPL_APPLIANCES="${DATADIR}/appliances.xml"
+export ARCHAPPL_MYIDENTITY=appliance0
 
 # Create an appliances.xml file and set up this appliance's identity.
-cat > appliances.xml <<EOF
+cat > "${ARCHAPPL_APPLIANCES}" <<EOF
  <appliances>
    <appliance>
      <identity>appliance0</identity>
@@ -71,11 +91,8 @@ cat > appliances.xml <<EOF
    </appliance>
  </appliances>
 EOF
- 
-export ARCHAPPL_APPLIANCES=${PWD}/appliances.xml
-export ARCHAPPL_MYIDENTITY=appliance0
 
-if [[ ! -f ${ARCHAPPL_APPLIANCES} ]]
+if [ ! -f "${ARCHAPPL_APPLIANCES}" ]
 then
 	echo "We just generated the appliances.xml file but we cannot seem to find it here ${ARCHAPPL_APPLIANCES}"
 	echo "The archiver appliance will not start successfully without this file"
@@ -86,17 +103,11 @@ fi
 # Set up the storage folders based on the default policy..
 export ARCHAPPL_SHORT_TERM_FOLDER=/dev/shm/quickstart_sts
 
-export ARCHAPPL_MEDIUM_TERM_FOLDER=${PWD}/MTS
-if [[ ! -d ${ARCHAPPL_MEDIUM_TERM_FOLDER} ]]
-then
-	mkdir ${ARCHAPPL_MEDIUM_TERM_FOLDER}
-fi
+export ARCHAPPL_MEDIUM_TERM_FOLDER="${DATADIR}/MTS"
+ensuredir "$ARCHAPPL_MEDIUM_TERM_FOLDER"
 
-export ARCHAPPL_LONG_TERM_FOLDER=${PWD}/MTS
-if [[ ! -d ${ARCHAPPL_LONG_TERM_FOLDER} ]]
-then
-	mkdir ${ARCHAPPL_LONG_TERM_FOLDER}
-fi
+export ARCHAPPL_LONG_TERM_FOLDER="${DATADIR}/LTS"
+ensuredir "$ARCHAPPL_LONG_TERM_FOLDER"
 
 # Use an in memory persistence layer, if one is not defined in the environment
 if [ -z ${ARCHAPPL_PERSISTENCE_LAYER} ] 
@@ -106,31 +117,32 @@ fi
 
 echo "Using ${ARCHAPPL_PERSISTENCE_LAYER} as the persistence layer"
 
-if [[ -d quickstart_tomcat ]]
+if [ -d quickstart_tomcat ]
 then
     echo "Found an older quickstart_tomcat folder. Stopping any existing instances."
-    pushd quickstart_tomcat
+    cd quickstart_tomcat
     for tomcatfolder in apache-tomcat*; do ${tomcatfolder}/bin/catalina.sh stop; done
     sleep 30
-    popd
-	echo "Deleting an older quickstart_tomcat"
-	rm -rf quickstart_tomcat
+    cd "$DATADIR"
+    echo "Deleting an older quickstart_tomcat"
+    rm -rf quickstart_tomcat
 fi
 
 mkdir quickstart_tomcat
-tar -C quickstart_tomcat -zxf  ${TOMCAT_DISTRIBUTION}
-pushd quickstart_tomcat
+tar -C quickstart_tomcat -zxf  "${TOMCAT_DISTRIBUTION}"
+cd quickstart_tomcat
+# eg. "apache-tomcat-7.0.57"
 TOMCAT_VERSION_FOLDER=`ls -d apache-tomca* | head -1`
 
 # Make sure we have a valid server.xml
-if [[ ! -f ${TOMCAT_VERSION_FOLDER}/conf/server.xml ]]
+if [ ! -f "${TOMCAT_VERSION_FOLDER}/conf/server.xml" ]
 then
   echo "After expanding the tomcat distribution, ${TOMCAT_DISTRIBUTION} into quickstart_tomcat/${TOMCAT_VERSION_FOLDER}, we cannot find ${TOMCAT_VERSION_FOLDER}/server.xml as expected."
   exit 1
 fi
 
 # Write a minimal server.xml into unpacked tomcat distribution.
-cat > ${TOMCAT_VERSION_FOLDER}/conf/server.xml <<EOF
+cat > "${TOMCAT_VERSION_FOLDER}/conf/server.xml" <<EOF
 <?xml version='1.0' encoding='utf-8'?>
 <!--
   Licensed to the Apache Software Foundation (ASF) under one or more
@@ -167,7 +179,7 @@ cat > ${TOMCAT_VERSION_FOLDER}/conf/server.xml <<EOF
 EOF
 
 # Write a log4.properties file into the lib folder
-cat > ${TOMCAT_VERSION_FOLDER}/lib/log4j.properties <<EOF
+cat > "${TOMCAT_VERSION_FOLDER}/lib/log4j.properties" <<EOF
 log4j.rootLogger=${DEFAULT_LOG_LEVEL}, console
 log4j.logger.org.apache.http=ERROR
 log4j.logger.config.org.epics.archiverappliance=DEBUG
@@ -178,35 +190,37 @@ log4j.appender.console.layout.ConversionPattern=%-4r [%t] %-5p %c %x - %m%n
 EOF
 
 # Now, deploy the WAR files. 
-cp ../mgmt.war ${TOMCAT_VERSION_FOLDER}/webapps
-cp ../engine.war ${TOMCAT_VERSION_FOLDER}/webapps
-cp ../etl.war ${TOMCAT_VERSION_FOLDER}/webapps
-cp ../retrieval.war ${TOMCAT_VERSION_FOLDER}/webapps
+cp "$SRCDIR/mgmt.war" ${TOMCAT_VERSION_FOLDER}/webapps
+cp "$SRCDIR/engine.war" ${TOMCAT_VERSION_FOLDER}/webapps
+cp "$SRCDIR/etl.war" ${TOMCAT_VERSION_FOLDER}/webapps
+cp "$SRCDIR/retrieval.war" ${TOMCAT_VERSION_FOLDER}/webapps
 
 # Unpack the mgmt war so that we can replace any images etc..
-pushd ${TOMCAT_VERSION_FOLDER}/webapps
-if [[ -d mgmt ]]
-then
-  rm -rf mgmt
-fi
+cd "${TOMCAT_VERSION_FOLDER}/webapps"
+
+[ -d mgmt ] && rm -rf mgmt
 mkdir mgmt
 cd mgmt
-jar xf ../mgmt.war
-popd
+jar xf "$SRCDIR/mgmt.war"
+
+cd "$DATADIR/quickstart_tomcat"
 
 pwd
 ls -ltr 
 
-if [[ -f ../site_specific_content/template_changes.html ]]
+if [ -f "$SRCDIR/site_specific_content/template_changes.html" ]
 then
   echo "Modifying static content to cater to site specific information"
-  java -cp ${TOMCAT_VERSION_FOLDER}/webapps/mgmt/WEB-INF/classes org.epics.archiverappliance.mgmt.bpl.SyncStaticContentHeadersFooters ../site_specific_content/template_changes.html ${TOMCAT_VERSION_FOLDER}/webapps/mgmt/ui
+  java -cp ${TOMCAT_VERSION_FOLDER}/webapps/mgmt/WEB-INF/classes \
+   org.epics.archiverappliance.mgmt.bpl.SyncStaticContentHeadersFooters \
+   "$SRCDIR/site_specific_content/template_changes.html" \
+   ${TOMCAT_VERSION_FOLDER}/webapps/mgmt/ui
 fi
 
-if [[ -d ../site_specific_content/img ]]
+if [ -d "$SRCDIR/site_specific_content/img" ]
 then
   echo "Replacing site specific images"
-  cp -R ../site_specific_content/img/* ${TOMCAT_VERSION_FOLDER}/webapps/mgmt/ui/comm/img/
+  cp -R "$SRCDIR/site_specific_content/img"/* ${TOMCAT_VERSION_FOLDER}/webapps/mgmt/ui/comm/img/
 fi
 
 
@@ -214,7 +228,7 @@ fi
 export JAVA_OPTS="-XX:MaxPermSize=128M -Xmx1G -ea"
 
 ARCH=`uname -m`
-if [[ $ARCH == 'x86_64' || $ARCH == 'amd64' ]]
+if [ "$ARCH" = 'x86_64' || "$ARCH" = 'amd64' ]
 then
   echo "Using 64 bit versions of libraries"
   export LD_LIBRARY_PATH=${TOMCAT_VERSION_FOLDER}/webapps/engine/WEB-INF/lib/native/linux-x86_64:${LD_LIBRARY_PATH}
@@ -224,7 +238,7 @@ else
 fi
 
 # Start tomcat up...
-if [[ $START_IN_BACKGROUND == "TRUE" ]]
+if [ "$START_IN_BACKGROUND" = "TRUE" ]
 then
   echo "Starting in the background..." 
   ${TOMCAT_VERSION_FOLDER}/bin/catalina.sh start
@@ -233,5 +247,3 @@ else
   echo "Starting in the foreground..." 
   ${TOMCAT_VERSION_FOLDER}/bin/catalina.sh run
 fi
-
-popd
